- greatly improved movement precision - bumped version to 1.2.1

+++ ThreeFingerDrag/mouse/cursor.h
@@ -5,7 +5,7 @@
 class Cursor {
 public:
 	static void SimulateClick(const DWORD click_flags);
-	static void MoveCursor(const int delta_x, const int delta_y);
+	static void MoveCursor(const double delta_x, const double delta_y);
 	static void LeftMouseDown();
 	static void LeftMouseUp();
 };
